refactor: 하이라이트 훅, 컴포넌트 폴더 구조 변경
