+++ frontend/src/components/highlight/components/HighlightEditor/style.ts
