+++ frontend/src/components/highlight/components/HighlightRemoverWrapper/index.tsx
