+++ frontend/src/components/highlight/components/Syntax/index.tsx
