+++ frontend/src/components/highlight/components/Syntax/style.ts
