+++ frontend/src/components/highlight/components/hooks/index.ts
