+++ frontend/src/components/highlight/components/hooks/useCheckHighlight.ts
