+++ frontend/src/components/highlight/components/hooks/useHighlight.ts
