+++ frontend/src/components/highlight/components/hooks/useHighlightRemoverPosition.ts
