+++ frontend/src/components/highlight/components/hooks/useHighlightToggleButtonPosition.ts
