+++ frontend/src/components/highlight/components/hooks/useMutateHighlight/index.ts
